Refactor esp32_ble_tracker to use esp32_ble core ble setup code
